% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.simpledb_operations.R
\name{domain_metadata}
\alias{domain_metadata}
\title{Returns information about the domain, including when the domain was created, the number of items and attributes in the domain, and the size of the attribute names and values}
\usage{
domain_metadata(
  DomainName = "string"
)
}
\arguments{
\item{DomainName}{[required] The name of the domain for which to display the metadata of.}
}
\description{
Returns information about the domain, including when the domain was created, the number of items and attributes in the domain, and the size of the attribute names and values.
}
\examples{

}
