% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.simpledb_package.R
\docType{package}
\name{paws.simpledb-package}
\alias{paws.simpledb}
\alias{paws.simpledb-package}
\title{paws.simpledb: Amazon SimpleDB}
\description{
Amazon SimpleDB is a web service providing the core database
functions of data indexing and querying in the cloud. By offloading
the time and effort associated with building and operating a web-scale
database, SimpleDB provides developers the freedom to focus on
application development. <p> A traditional, clustered relational
database requires a sizable upfront capital outlay, is complex to
design, and often requires extensive and repetitive database
administration. Amazon SimpleDB is dramatically simpler, requiring no
schema, automatically indexing your data and providing a simple API
for storage and access. This approach eliminates the administrative
burden of data modeling, index maintenance, and performance tuning.
Developers gain access to this functionality within Amazon's proven
computing environment, are able to scale instantly, and pay only for
what they use. </p> <p> Visit <a
href="http://aws.amazon.com/simpledb/">http://aws.amazon.com/simpledb/</a>
for more information. </p>
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
